Remove parallelism code from util.py, kill util.py
Util.py is now gone, in favor of having categories for various util
functionality.

This commit also changes some import syntax, preferring to import "foo"
and invoke "foo.SomeMethod", rather than directly importing the method,
e.g. "from gslib.foo import SomeMethod; SomeMethod()".  We should
eventually do this throughout the entire code base (in another commit),
but I only changed it in the files where it was a simple fix.
