+++ gslib/command_runner.py
@@ -43,12 +43,12 @@ from gslib.exception import CommandException
 from gslib.gcs_json_api import GcsJsonApi
 from gslib.no_op_credentials import NoOpCredentials
 from gslib.tab_complete import MakeCompleter
-from gslib.util import CheckMultiprocessingAvailableAndInit
 from gslib.utils import boto_util
 from gslib.utils.constants import GSUTIL_PUB_TARBALL
 from gslib.utils.constants import RELEASE_NOTES_URL
 from gslib.utils.constants import UTF8
 from gslib.utils.metadata_util import IsCustomMetadataHeader
+from gslib.utils.parallelism_framework_util import CheckMultiprocessingAvailableAndInit
 from gslib.utils.system_util import IsRunningInteractively
 from gslib.utils.text_util import CompareVersions
 from gslib.utils.text_util import InsistAsciiHeader
