+++ gslib/commands/hash.py
@@ -31,14 +31,16 @@ from gslib.hashing_helper import Base64EncodeHash
 from gslib.hashing_helper import Base64ToHexHash
 from gslib.hashing_helper import CalculateHashesFromContents
 from gslib.hashing_helper import SLOW_CRCMOD_WARNING
-from gslib.parallelism_framework_util import PutToQueueWithTimeout
 from gslib.progress_callback import FileProgressCallbackHandler
 from gslib.progress_callback import ProgressCallbackWithTimeout
 from gslib.storage_url import StorageUrlFromString
 from gslib.thread_message import FileMessage
 from gslib.thread_message import FinalMessage
-from gslib.utils.boto_util import UsingCrcmodExtension
-from gslib.utils.constants import NO_MAX
+from gslib.utils import boto_util
+from gslib.utils import constants
+from gslib.utils import parallelism_framework_util
+
+_PutToQueueWithTimeout = parallelism_framework_util.PutToQueueWithTimeout
 
 _SYNOPSIS = """
   gsutil hash [-c] [-h] [-m] filename...
@@ -79,7 +81,7 @@ class HashCommand(Command):
       command_name_aliases=[],
       usage_synopsis=_SYNOPSIS,
       min_args=1,
-      max_args=NO_MAX,
+      max_args=constants.NO_MAX,
       supported_sub_args='chm',
       file_url_ok=True,
       provider_url_ok=False,
@@ -141,7 +143,7 @@ class HashCommand(Command):
       calc_crc32c = True
       calc_md5 = True
 
-    if calc_crc32c and not UsingCrcmodExtension(crcmod):
+    if calc_crc32c and not boto_util.UsingCrcmodExtension(crcmod):
       logger.warn(SLOW_CRCMOD_WARNING)
 
     return calc_crc32c, calc_md5, format_func, cloud_format_func, output_format
@@ -216,7 +218,7 @@ class HashCommand(Command):
 
     if not matched_one:
       raise CommandException('No files matched')
-    PutToQueueWithTimeout(self.gsutil_api.status_queue,
+    _PutToQueueWithTimeout(self.gsutil_api.status_queue,
                           FinalMessage(time.time()))
     return 0
