+++ gslib/commands/setmeta.py
@@ -29,18 +29,20 @@ from gslib.cs_api_map import ApiSelector
 from gslib.exception import CommandException
 from gslib.name_expansion import NameExpansionIterator
 from gslib.name_expansion import SeekAheadNameExpansionIterator
-from gslib.parallelism_framework_util import PutToQueueWithTimeout
 from gslib.storage_url import StorageUrlFromString
 from gslib.third_party.storage_apitools import storage_v1_messages as apitools_messages
 from gslib.thread_message import MetadataMessage
 from gslib.translation_helper import CopyObjectMetadata
 from gslib.translation_helper import ObjectMetadataFromHeaders
 from gslib.translation_helper import PreconditionsFromHeaders
-from gslib.utils.constants import NO_MAX
+from gslib.utils import constants
+from gslib.utils import parallelism_framework_util
 from gslib.utils.metadata_util import IsCustomMetadataHeader
 from gslib.utils.retry_util import Retry
 from gslib.utils.text_util import InsistAsciiHeader
 from gslib.utils.text_util import InsistAsciiHeaderValue
+
+_PutToQueueWithTimeout = parallelism_framework_util.PutToQueueWithTimeout
 
 _SYNOPSIS = """
   gsutil setmeta -h [header:value|header] ... url...
@@ -137,7 +139,7 @@ class SetMetaCommand(Command):
       command_name_aliases=['setheader'],
       usage_synopsis=_SYNOPSIS,
       min_args=1,
-      max_args=NO_MAX,
+      max_args=constants.NO_MAX,
       supported_sub_args='h:rR',
       file_url_ok=False,
       provider_url_ok=False,
@@ -259,7 +261,7 @@ class SetMetaCommand(Command):
         exp_src_url.bucket_name, exp_src_url.object_name, patch_obj_metadata,
         generation=exp_src_url.generation, preconditions=preconditions,
         provider=exp_src_url.scheme, fields=['id'])
-    PutToQueueWithTimeout(gsutil_api.status_queue,
+    _PutToQueueWithTimeout(gsutil_api.status_queue,
                           MetadataMessage(message_time=time.time()))
 
   def _ParseMetadataHeaders(self, headers):
