+++ gslib/commands/version.py
@@ -26,9 +26,9 @@ import boto
 import crcmod
 import gslib
 from gslib.command import Command
-from gslib.util import CheckMultiprocessingAvailableAndInit
 from gslib.utils.boto_util import GetConfigFilePaths
 from gslib.utils.boto_util import UsingCrcmodExtension
+from gslib.utils.parallelism_framework_util import CheckMultiprocessingAvailableAndInit
 
 
 _SYNOPSIS = """
