+++ gslib/daisy_chain_wrapper.py
@@ -23,8 +23,8 @@ import time
 from gslib.cloud_api import BadRequestException
 from gslib.cloud_api import CloudApi
 from gslib.encryption_helper import CryptoKeyWrapperFromKey
-from gslib.util import CreateLock
-from gslib.utils.constants import TRANSFER_BUFFER_SIZE
+from gslib.utils import constants
+from gslib.utils import parallelism_framework_util
 
 
 # This controls the amount of bytes downloaded per download request.
@@ -79,8 +79,8 @@ class DaisyChainWrapper(object):
   to copy a file.
 
   This class is coupled with the XML and JSON implementations in that it
-  expects that small buffers (maximum of TRANSFER_BUFFER_SIZE) in size will be
-  used.
+  expects that small buffers (maximum of constants.TRANSFER_BUFFER_SIZE) in
+  size will be used.
   """
 
   def __init__(self, src_url, src_obj_size, gsutil_api,
@@ -121,10 +121,10 @@ class DaisyChainWrapper(object):
     self.last_data = None
 
     # Protects buffer, position, bytes_buffered, last_position, and last_data.
-    self.lock = CreateLock()
+    self.lock = parallelism_framework_util.CreateLock()
 
     # Protects download_exception.
-    self.download_exception_lock = CreateLock()
+    self.download_exception_lock = parallelism_framework_util.CreateLock()
 
     self.src_obj_size = src_obj_size
     self.src_url = src_url
@@ -212,10 +212,10 @@ class DaisyChainWrapper(object):
       # If there is no data left or 0 bytes were requested, return an empty
       # string so callers can call still call len() and read(0).
       return ''
-    if amt is None or amt > TRANSFER_BUFFER_SIZE:
+    if amt is None or amt > constants.TRANSFER_BUFFER_SIZE:
       raise BadRequestException(
           'Invalid HTTP read size %s during daisy chain operation, '
-          'expected <= %s.' % (amt, TRANSFER_BUFFER_SIZE))
+          'expected <= %s.' % (amt, constants.TRANSFER_BUFFER_SIZE))
 
     while True:
       with self.lock:
