+++ gslib/tests/test_ui.py
@@ -27,17 +27,16 @@ from __future__ import absolute_import
 from hashlib import md5
 import os
 import pickle
-import Queue
 import StringIO
 
 import crcmod
+from six.moves import queue as Queue
+
 from gslib.copy_helper import PARALLEL_UPLOAD_STATIC_SALT
 from gslib.copy_helper import PARALLEL_UPLOAD_TEMP_NAMESPACE
 from gslib.cs_api_map import ApiSelector
 from gslib.parallel_tracker_file import ObjectFromTracker
 from gslib.parallel_tracker_file import WriteParallelUploadTrackerFile
-from gslib.parallelism_framework_util import PutToQueueWithTimeout
-from gslib.parallelism_framework_util import ZERO_TASKS_TO_DO_ARGUMENT
 from gslib.storage_url import StorageUrlFromString
 import gslib.tests.testcase as testcase
 from gslib.tests.testcase.integration_testcase import SkipForS3
@@ -68,6 +67,8 @@ from gslib.ui_controller import UIThread
 from gslib.utils.boto_util import UsingCrcmodExtension
 from gslib.utils.constants import START_CALLBACK_PER_BYTES
 from gslib.utils.constants import UTF8
+from gslib.utils.parallelism_framework_util import PutToQueueWithTimeout
+from gslib.utils.parallelism_framework_util import ZERO_TASKS_TO_DO_ARGUMENT
 from gslib.utils.retry_util import Retry
 from gslib.utils.unit_util import HumanReadableWithDecimalPlaces
 from gslib.utils.unit_util import MakeHumanReadable
